requirements.txt
torch>=2.1.0
numpy>=1.21.0
PyYAML>=6.0
matplotlib>=3.5.0
tensorboard>=2.10.0
tqdm>=4.64.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.3.0
pytest>=7.0.0
